requirements.txt
pymongo
python-dotenv
opencv-python